requirements.txt
Django==5.2.10
djangorestframework==3.16.1
django-filter==25.1
django-simple-history==3.10.1
django-import-export==4.1.1
Pillow==12.1.0
flake8==7.1.1
